requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnevnikru==1.3.3
docopt==0.6.2
ffmpeg==1.4
Flask==2.2.3
fluent-logger==0.10.0
frozenlist==1.3.3
google-search-results==2.4.1
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
msgpack==1.0.4
multidict==6.0.4
openai==0.27.0
pipreqs==0.4.11
pydantic==1.10.2
pydub==0.25.1
pyee==9.0.4
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
python-ffmpeg==2.0.2
requests==2.28.1
semver==2.13.0
soupsieve==2.4
steamship==2.13.16
telebot==0.0.5
toml==0.10.2
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
yarg==0.1.9
yarl==1.8.2
